backend/requirements.txt
livekit
livekit-agents
livekit-api
mem0ai
google-generativeai
fastapi
uvicorn
python-dotenv
google.genai